kafka: add CommonConfig
Refactoring to avoid duplicating config
in ProducerConfig and ConsumerConfig.
